seller order history

+++ app/src/main/java/org/d3if2101/canteen/ui/penjual/dashboard/DashboardPenjualActivity.kt
@@ -1,5 +1,6 @@
 package org.d3if2101.canteen.ui.penjual.dashboard
 
+import android.app.AlertDialog
 import android.content.Intent
 import android.os.Bundle
 import android.os.Handler
@@ -13,6 +14,7 @@ import androidx.core.view.GravityCompat
 import androidx.drawerlayout.widget.DrawerLayout
 import com.google.android.material.navigation.NavigationView
 import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.auth.ktx.auth
 import com.google.firebase.ktx.Firebase
 import org.d3if2101.canteen.R
 import org.d3if2101.canteen.databinding.ActivityDashboardPenjualBinding
@@ -102,7 +104,7 @@ class DashboardPenjualActivity : AppCompatActivity() {
                 }
                 R.id.nav_log_out -> {
                     drawerLayout.closeDrawer(GravityCompat.START)
-                    logout()
+                    logOutUser()
                 }
             }
             true
@@ -117,11 +119,24 @@ class DashboardPenjualActivity : AppCompatActivity() {
         }
     }
 
-    private fun logout() {
-        val intent = Intent(this, Login::class.java)
-        startActivity(intent)
-        finish()
-        viewModel.deleteTokenPref()
-        Toast.makeText(this, "Logged out", Toast.LENGTH_SHORT).show()
+    private fun logOutUser() {
+        AlertDialog.Builder(this)
+            .setTitle("Peringatan")
+            .setMessage("Apa kamu yakin untuk Log Out?")
+            .setPositiveButton("Ya") { _, _ ->
+                Firebase.auth.signOut()
+
+                getSharedPreferences("settings", MODE_PRIVATE).edit().clear().apply()
+                getSharedPreferences("user_profile_details", MODE_PRIVATE).edit().clear().apply()
+
+                viewModel.deleteTokenPref()
+                Toast.makeText(this, "Logged out", Toast.LENGTH_SHORT).show()
+
+                startActivity(Intent(this, Login::class.java))
+                finish()
+            }
+            .setNegativeButton("Tidak") { dialogInterface, _ ->
+                dialogInterface.dismiss()
+            }.create().show()
     }
 }
